refactor(arm64): split codegen into focused modules
