+++ components/aihc-arm64/src/Aihc/Arm64/Codegen/Snapshot.hs
@@ -1,0 +1,163 @@
+{-# LANGUAGE OverloadedStrings #-}
+
+-- | Render the C metadata consumed by observed-program snapshots.
+module Aihc.Arm64.Codegen.Snapshot (renderObservedMetadata) where
+
+import Aihc.Arm64.Codegen.Runtime (functionCodeLabel, tshow)
+import Aihc.Arm64.Codegen.Types
+import Aihc.Grin.Syntax
+import Aihc.Tc.Types (RuntimeRep (..))
+import Data.Map.Strict qualified as Map
+import Data.Text (Text)
+import Data.Text qualified as T
+
+renderObservedMetadata :: CompileEnv -> GrinProgram -> [RuntimeRep] -> Either Arm64Error Text
+renderObservedMetadata env program resultReps = do
+  renderedResultReps <- mapM snapshotRepName resultReps
+  constructors <- mapM renderConstructorDescriptor constructorEntries
+  functions <- mapM renderFunctionDescriptor functionEntries
+  pure . T.unlines $
+    [ "#include \"aihc_runtime.h\"",
+      "#include <stddef.h>",
+      ""
+    ]
+      <> map renderFunctionDeclaration functions
+      <> [""]
+      <> concatMap renderConstructorRepDeclaration constructors
+      <> concatMap renderFunctionRepDeclaration functions
+      <> renderRepDeclaration "result_reps" renderedResultReps
+      <> renderConstructorTable constructors
+      <> renderFunctionTable functions
+      <> [ "void aihc_snapshot_dump_result(uint64_t count, const AihcSlot *values, const AihcMachine *machine) {",
+           "  aihc_snapshot_dump(count, values, " <> pointerOrNull renderedResultReps "result_reps" <> ",",
+           "                     aihc_allocation_count(machine),",
+           "                     " <> tshow (length constructors) <> ", " <> tableOrNull constructors "constructors" <> ",",
+           "                     " <> tshow (length functions) <> ", " <> tableOrNull functions "functions" <> ");",
+           "}"
+         ]
+  where
+    layouts =
+      Map.fromList
+        ( builtinConstructorLayouts
+            <> [(name, concat argumentLayouts) | (name, argumentLayouts) <- grinConstructors program]
+        )
+    constructorEntries =
+      [ (identifier, name, fields)
+      | (name, identifier) <- Map.toAscList (compileConstructorIds env),
+        Just fields <- [Map.lookup name layouts]
+      ]
+    localFunctionEntries =
+      [ (grinFunctionName function, map grinVarRuntimeRep (grinFunctionParameters function))
+      | function <- grinFunctions program
+      ]
+    externalFunctionEntries =
+      [ (grinCodeFunctionName info, concat (grinCodeParameterLayouts info))
+      | info <- grinExternalFunctions program
+      ]
+    functionEntries = externalFunctionEntries <> localFunctionEntries
+
+    renderConstructorDescriptor (identifier, name, fields) = do
+      reps <- mapM snapshotRepName fields
+      pure (identifier, name, reps)
+
+    renderFunctionDescriptor (name, parameters) = do
+      label <- functionCodeLabel env name
+      reps <- mapM snapshotRepName parameters
+      pure (name, label, reps)
+
+renderFunctionDeclaration :: (FunctionName, Text, [Text]) -> Text
+renderFunctionDeclaration (_, label, _) =
+  "extern void " <> cSymbol label <> "(void);"
+
+renderConstructorRepDeclaration :: (Int, Text, [Text]) -> [Text]
+renderConstructorRepDeclaration (identifier, _, reps) =
+  renderRepDeclaration ("constructor_reps_" <> tshow identifier) reps
+
+renderFunctionRepDeclaration :: (FunctionName, Text, [Text]) -> [Text]
+renderFunctionRepDeclaration (_, label, reps) =
+  renderRepDeclaration ("function_reps_" <> cSymbol label) reps
+
+renderRepDeclaration :: Text -> [Text] -> [Text]
+renderRepDeclaration _ [] = []
+renderRepDeclaration name reps =
+  [ "static const AihcSnapshotRep "
+      <> name
+      <> "[] = {"
+      <> T.intercalate ", " reps
+      <> "};"
+  ]
+
+renderConstructorTable :: [(Int, Text, [Text])] -> [Text]
+renderConstructorTable [] = []
+renderConstructorTable constructors =
+  ["static const AihcSnapshotConstructor constructors[] = {"]
+    <> [ "  {"
+           <> tshow identifier
+           <> ", "
+           <> cString name
+           <> ", "
+           <> tshow (length reps)
+           <> ", "
+           <> pointerOrNull reps ("constructor_reps_" <> tshow identifier)
+           <> "},"
+       | (identifier, name, reps) <- constructors
+       ]
+    <> ["};"]
+
+renderFunctionTable :: [(FunctionName, Text, [Text])] -> [Text]
+renderFunctionTable [] = []
+renderFunctionTable functions =
+  ["static const AihcSnapshotFunction functions[] = {"]
+    <> [ "  {(uintptr_t)&"
+           <> cSymbol label
+           <> ", "
+           <> cString (unFunctionName name)
+           <> ", "
+           <> tshow (length reps)
+           <> ", "
+           <> pointerOrNull reps ("function_reps_" <> cSymbol label)
+           <> "},"
+       | (name, label, reps) <- functions
+       ]
+    <> ["};"]
+
+snapshotRepName :: RuntimeRep -> Either Arm64Error Text
+snapshotRepName runtimeRep =
+  case runtimeRep of
+    BoxedRep {} -> pure "AIHC_SNAPSHOT_POINTER"
+    SumRep {} -> pure "AIHC_SNAPSHOT_POINTER"
+    IntRep -> pure "AIHC_SNAPSHOT_INT"
+    Int8Rep -> pure "AIHC_SNAPSHOT_INT8"
+    Int16Rep -> pure "AIHC_SNAPSHOT_INT16"
+    Int32Rep -> pure "AIHC_SNAPSHOT_INT32"
+    Int64Rep -> pure "AIHC_SNAPSHOT_INT64"
+    WordRep -> pure "AIHC_SNAPSHOT_WORD"
+    Word8Rep -> pure "AIHC_SNAPSHOT_WORD8"
+    Word16Rep -> pure "AIHC_SNAPSHOT_WORD16"
+    Word32Rep -> pure "AIHC_SNAPSHOT_WORD32"
+    Word64Rep -> pure "AIHC_SNAPSHOT_WORD64"
+    AddrRep -> pure "AIHC_SNAPSHOT_ADDR"
+    FloatRep -> pure "AIHC_SNAPSHOT_FLOAT"
+    DoubleRep -> pure "AIHC_SNAPSHOT_DOUBLE"
+    _ -> Left (Arm64UnsupportedRuntimeRep runtimeRep)
+
+pointerOrNull :: [value] -> Text -> Text
+pointerOrNull values name
+  | null values = "NULL"
+  | otherwise = name
+
+tableOrNull :: [value] -> Text -> Text
+tableOrNull = pointerOrNull
+
+cSymbol :: Text -> Text
+cSymbol = T.drop 1
+
+cString :: Text -> Text
+cString value = "\"" <> T.concatMap escape value <> "\""
+  where
+    escape '"' = "\\\""
+    escape '\\' = "\\\\"
+    escape '\n' = "\\n"
+    escape '\r' = "\\r"
+    escape '\t' = "\\t"
+    escape character = T.singleton character
